implemented adjusted class and adjusted utils method
